+++ src/domain/song/entity.rs
@@ -1,10 +1,10 @@
 #[allow(dead_code)]
 #[derive(sqlx::FromRow, Debug, Clone)]
 pub struct Song {
-    pub id: i64,
+    pub id: i32,
     pub title: String,
     pub artist: String,
-    pub release_year: i64,
+    pub release_year: i16,
     pub album: String,
     pub remix: String,
     pub search_blob: String,
