+++ src/domain/song_filter/entity.rs
@@ -1,7 +1,7 @@
 #[allow(dead_code)]
 #[derive(sqlx::FromRow, Debug, Clone)]
 pub struct SongFilter {
-    pub id: i64,
-    pub song_id: i64,
-    pub filter_id: i64,
+    pub id: i32,
+    pub song_id: i32,
+    pub filter_id: i32,
 }
